Move page components to containers
