+++ app/pages/[eventId]/books/[id]/edit.ts
@@ -1,0 +1,2 @@
+import page from '../../../../containers/BookEdit'
+export default page
