+++ app/pages/[eventId]/books/new.ts
@@ -1,0 +1,2 @@
+import page from '../../../containers/BookNew'
+export default page
