+++ app/pages/[eventId]/circles/[id]/edit.ts
@@ -1,0 +1,2 @@
+import page from '../../../../containers/CircleEdit'
+export default page
